[All] Give a init() method to every module (2)
