+++ SofaKernel/modules/SofaSimulationGraph/CMakeLists.txt
@@ -6,6 +6,7 @@ set(SOFASIMULATIONGRAPH_SRC "src/${PROJECT_NAME}")
 set(HEADER_FILES
     ${SOFASIMULATIONGRAPH_SRC}/config.h.in
     ${SOFASIMULATIONGRAPH_SRC}/init.h
+    ${SOFASIMULATIONGRAPH_SRC}/initSofaSimulationGraph.h
     ${SOFASIMULATIONGRAPH_SRC}/graph.h
     ${SOFASIMULATIONGRAPH_SRC}/DAGNode.h
     ${SOFASIMULATIONGRAPH_SRC}/DAGNodeMultiMappingElement.h
@@ -15,6 +16,7 @@ set(HEADER_FILES
 
 set(SOURCE_FILES
     ${SOFASIMULATIONGRAPH_SRC}/init.cpp
+    ${SOFASIMULATIONGRAPH_SRC}/initSofaSimulationGraph.cpp
     ${SOFASIMULATIONGRAPH_SRC}/DAGNode.cpp
     ${SOFASIMULATIONGRAPH_SRC}/DAGNodeMultiMappingElement.cpp
     ${SOFASIMULATIONGRAPH_SRC}/DAGSimulation.cpp
