+++ modules/SofaPreconditioner/src/SofaPreconditioner/initSofaPreconditioner.h
@@ -19,59 +19,11 @@
 *                                                                             *
 * Contact information: contact@sofa-framework.org                             *
 ******************************************************************************/
-#include <SofaValidation/initValidation.h>
-
-#include <sofa/core/ObjectFactory.h>
-using sofa::core::ObjectFactory;
+#include <SofaPreconditioner/config.h>
 
 namespace sofa::component
 {
 
+SOFA_PRECONDITIONER_API void initSofaPreconditioner();
 
-extern "C" {
-    SOFA_SOFAVALIDATION_API void initExternalModule();
-    SOFA_SOFAVALIDATION_API const char* getModuleName();
-    SOFA_SOFAVALIDATION_API const char* getModuleVersion();
-    SOFA_SOFAVALIDATION_API const char* getModuleLicense();
-    SOFA_SOFAVALIDATION_API const char* getModuleDescription();
-    SOFA_SOFAVALIDATION_API const char* getModuleComponentList();
 }
-
-void initExternalModule()
-{
-    static bool first = true;
-    if (first)
-    {
-        first = false;
-    }
-}
-
-const char* getModuleName()
-{
-    return sofa_tostring(SOFA_TARGET);
-}
-
-const char* getModuleVersion()
-{
-    return sofa_tostring(SOFAVALIDATION_VERSION);
-}
-
-const char* getModuleLicense()
-{
-    return "LGPL";
-}
-
-const char* getModuleDescription()
-{
-    return "This plugin contains utilities used for validation purpose.";
-}
-
-const char* getModuleComponentList()
-{
-    /// string containing the names of the classes provided by the plugin
-    static std::string classes = ObjectFactory::getInstance()->listClassesFromTarget(sofa_tostring(SOFA_TARGET));
-    return classes.c_str();
-}
-
-} // namespace sofa::component
-
