+++ modules/SofaValidation/src/SofaValidation/initSofaValidation.cpp
@@ -19,9 +19,64 @@
 *                                                                             *
 * Contact information: contact@sofa-framework.org                             *
 ******************************************************************************/
-#ifndef SOFA_COMPONENT_EXPORTER_INIT_H
-#define SOFA_COMPONENT_EXPORTER_INIT_H
-#include <SofaExporter/config.h>
+#include <SofaValidation/initSofaValidation.h>
 
-#endif
+#include <sofa/core/ObjectFactory.h>
+using sofa::core::ObjectFactory;
+
+namespace sofa::component
+{
+
+
+void initSofaValidation()
+{
+    static bool first = true;
+    if (first)
+    {
+        first = false;
+    }
+}
+
+extern "C" {
+    SOFA_SOFAVALIDATION_API void initExternalModule();
+    SOFA_SOFAVALIDATION_API const char* getModuleName();
+    SOFA_SOFAVALIDATION_API const char* getModuleVersion();
+    SOFA_SOFAVALIDATION_API const char* getModuleLicense();
+    SOFA_SOFAVALIDATION_API const char* getModuleDescription();
+    SOFA_SOFAVALIDATION_API const char* getModuleComponentList();
+}
+
+void initExternalModule()
+{
+    initSofaValidation();
+}
+
+const char* getModuleName()
+{
+    return sofa_tostring(SOFA_TARGET);
+}
+
+const char* getModuleVersion()
+{
+    return sofa_tostring(SOFAVALIDATION_VERSION);
+}
+
+const char* getModuleLicense()
+{
+    return "LGPL";
+}
+
+const char* getModuleDescription()
+{
+    return "This plugin contains utilities used for validation purpose.";
+}
+
+const char* getModuleComponentList()
+{
+    /// string containing the names of the classes provided by the plugin
+    static std::string classes = ObjectFactory::getInstance()->listClassesFromTarget(sofa_tostring(SOFA_TARGET));
+    return classes.c_str();
+}
+
+} // namespace sofa::component
